+++ packages/types/src/main/IAnyMap.ts
@@ -1,0 +1,4 @@
+export type IAnyMap = {
+  [key: string]: any
+  [key: number]: any
+}